reg_usage: Enhanced instruction tables with register usage data
This is the first version of "reg_usage" example demonstrating the ways to
find out which registers a given instruction uses.

This version includes the enhanced instruction tables, modified helper
script (gen-insn-attr-x86.awk) and the decoder. The latter is almost the
same as used in KProbes and KEDR.

Currently, it is possible to only analyze the opcode and find out the
corresponding part of register usage data. The future versions of
"reg_usage" should be able to extract the remaining parts of these data from
ModRM and SIB bytes of an instruction as well as detect commonly used no-op
instructions of several kinds.