reg_usage: Enhanced the decoder to provide more data for each instruction
1. The instruction attributes are now contained in a struct rather than a
plain unsigned int. The decoder was enhanced to correctly deal with that.
This allows extracting more information about the instructions from their
specification in the opcode table.

Based on this enhancement, it is planned to implement functions providing
data about the types of the operands, addressing methods used by these
operands as well as memory access possibility.

2. Provided a stub for insn_register_usage_mask(), the function that
returns register usage information for a given instruction. This time, it is
not only the opcode but also the other parts of an instruction that are to
be taken into account. The function is still to be implemented.