reg_usage: Register usage masks for ModRM.RM & SIB are available separately
insn_reg_mask_for_expr() has been prepared. It finds out which registers are
used for ModRM-style memory addressing. The function analyzes ModRM.RM and
SIB fields only.

To have the means to distinguish which registers are used by an instruction
as a whole and which ones are used in memory addressing expression only is a
convenient thing. It allows, for example, to lower the "register pressure"
to some extent when allocating registers for the instrumentation code.