i13n: Prepared a skeleton of "Instrumentation" example
"Instrumentation" example should eventually become a prototype of the
instrumentation subsystem of Kernel Strider.

For now it is just a skeleton: stubs are there instead of the actual
instrumentation facilities. Still, several important components are already
in place: code relocation mechanism, fallback function support, jump table
processing, resource allocation mechanism, etc.

Currently, the example works as follows. For each function from the target
module, the system creates a "fallback" instance, i.e. a relocated copy of
the code, and an instrumented instance. The latter is just a stub now, a
jump to the fallback instance. The first 5 bytes of the original function
are overwritten with a jump to the instrumented instance. The target module
must remain usable after such surgery.

In the future, the fallback instances will be used if the system
fails to allocate "thread-local" memory.