i13n: Refactored "Instrumentation" example
function.c was becoming overweight, so a kind of a functional decomposition
has been performed. Utility stuff that could be used in several places went
to util.*, everything related to the creation of the instrumented instance
of a function - to instrument.*, etc.

In addition, the Makefile was modified to properly handle dependency
changes. Before this, the changes in the instruction analysis subsystem did
not affect the build process and explicit 'make clean' was needed.