i13n: Provided definitions of primary storage, operations with it, etc.
Some functions are actually stubs now but the definitions of struct
kedr_primary_storage and the auxiliary structures are now in place. So are
the definitions of the wrappers for the operations to be used from the
instrumented code as well as of their holder functions.

The operations themselves are to be implemented later.