i13n: Relocation mechanism has been rewritten
The previous variant of relocation mechanism (used for jmp/call rel32 and
for the instructions with RIP-relative addressing) was inconsistent and
buggy (esp. relocating to the instruction address of 0).

Now the system maintains the list of the instructions to be relocated during
the deployment phase of an instrumented function and should perform the
relocation properly.

Apart from that, many minor fixes and enhancements have been made.