i13n: Added checking for "split" functions; added processing of short jumps
Incomplete ("split") functions may appear if there are '.global' or '.local'
symbol definitions in the inline assembly  within an original function. As a
result, the system will see 2 or more separate functions instead of one.
Our system currently cannot handle such situations, it should issue a
warning and refuse to instrument the target module.

At this stage, we detect common (but still not all) situations of this kind.

Note that, the issues with ".altinstr_replacement" section are not handled
by the detector of incomplete functions yet (for the present, that section
is erroneously considered to be a part of the previous function).