sections: Improved handling of bad input in the parser
The items in the buffer to be parsed should be separated by spaces
and/or tabs. In some cases, particularly for the input like
".txt 0x01234.init 098" (note there is no space before ".init") the
parser considered this input valid although it is not.

The problem is fixed by this patch.