i13n: Fixed calculation of function sizes (taking sections into account)
On some systems (e.g. on Fedora 15 x86), the sections containing the
functions to be instrumented (.text, .init.text, ...) may have other
sections inbetween (.altinstr_replacement, etc.). The mechanism we used to
determine the size of each function did not take this into account and could
report sizes greater than they actually are. This could result in many nasty
problems.

Now the addresses of all the loaded sections of a module are obtained and
this information is used to produce a correct estimate of the size of each
function. The technique was taken from "Sections" example.