i13n: Prepared the skeleton of phase #1 of IR transformation
The first series of transformations of the IR deals with the special
constructs in the code that should be handled (function entry, function
exit, end of a block, ...) as well as processes the instructions that use
the base register.

To implement this phase, we need to replace the stubs of the handlers for
the particular constructs with real handlers.

Several cosmetic changes have been made as well.