i13n: Revisited handling of jumps out of block; added processing of XLAT
1. The jumps out of a normal block are now handled at phase 2 of IR
transformations and only if the block has memory accesses of interest. Same
for handling of block end: don't bother if there are no memory accesses
there. The primary storage should be in the required state anyway.

2. Fixed: memory accesses from XLAT insn were not taken into account.
Prepared a stub of a handler for XLAT, added detection of this insn at the
phase 2.