i13n: Implemented phase 1 of IR transformation
The following constructs in the code are instrumented during the 1st phase
of IR transformation:
- function entry and exits;
- indirect near jumps and calls;
- the instructions that use the chosen base register (%base, the register
that points to the "thread local" storage for the collected data).

The handlers for these constructs are now in place. A refactoring of some
other subsystems was also performed to be able to prepare these handlers.