i13n: Implemented processing of block ends and jumps out of the blocks
The handlers for the ends of the blocks as well as for the inner forward
jumps out of the blocks have been added.

Some other things were cleaned up, added or fixed during the process. For
example, support for another type of relocations was added. It is also
necessary to handle jumps out of the blocks properly. Besides that:
- a more generic generator of 'mov %reg,<offset>(%base)' and the like has
  been added;
- the implementation of the generator of 'or <value>,<offset>(%base)' has
  been revisited to create smaller instructions in case of short blocks;
- etc.