i13n: Added processing of memory accesses, types: E, M; made other improvements
The memory accesses from the instructions of types E and M are now
processed. These types are the most common among the instructions operating
on memory - see Intel's manual vol. 2B, "A.2.1 Codes for Addressing Method",
for details.

Besides that, the following improvements have been made.
1.
Processing of the memory accesses relative to %rsp/%esp is now off by
default. Module parameter "process_sp_accesses" controls this. When it is 0,
such memory accesses are not instrumented at all (although there is still
room for their records in the primary storage, for simplicity). Less code
bloat as a result. In addition, such memory operations usually deal with
the stack only, i.e. with the thread-local data. This is not always
interesting to register, especially in hunting down data races. Note that
this does not filter out ALL stack accesses. Address filtering is desirable
for that and it should be implemented in the future.

2.
Handling of block ends has been improved in order not to generate unneeded
operations. For example, 'lock_mask' is 0 on entry to the handler, so if
the block contained no locked operations, it is not needed to generate
'mov 0, ps->lock_mask'. Similar things - for ORing read and write masks.