i13n: Fixed: MOVSXD was assumed to access a quadword on x86-64
In Intel's manual vol 2B, MOVSXD instruction (opcode 63/r) is presented as
follows in Table A-2: MOVSXD (o64) Gv, Ev.
In our example, the decoder assumes the same parameters of this instruction.

The type of the argument that refers to a memory location (Ev) seems to be
incorrect. According to section "A.2.2 Codes for Operand Type" of that
manual, "v" means:

   "Word, doubleword or quadword (in 64-bit mode), depending on operand-size
attribute."

MOVSXD, however, always accesses a doubleword in memory, so the correct type
should probably be "Ed".

This problem resulted in recording the access to 8 bytes rather than 4 on
x86-64 systems. It should be fixed by this commit.