i13n: Added handling of direct offset MOVs
Memory accesses made by direct offset MOV instructions (opcodes: A0-A3) are
now also recorded.

This kind of MOV is sometimes used in the kernel modules on 32-bit systems
to load the value of some global or static variable in %eax or %al or to
store the value of %eax or %al in such variable.

Direct offset MOVs seem to be used more rarely on x86-64. Still, we track
these on 64-bit systems too.