core: The owner of the handlers is now locked in memory when needed
When the target has just loaded, the core now increases the usage count
of the module that owns the currently registered set of event handlers.
That module becomes not unloadable until the target is unloaded.

The core module is now also not unloadable when the target is in memory:
either the core is the owner of the current set of handlers or the owner
of that set is another module that uses the functions exported by the
core. The usage count of the core module will be positive in both cases
this way.