core: Added section subsystem from "Instrumentation" example
The section subsystem from "Instrumentation" was added to the core with
several modifications. The most important one is that the parameters of the
subsystem (path to the helper, name of the file in debugfs, etc.) are now
configurable by the build system.

The build system was revisited to accomodate these changes: the helper
script that obtains section names and addresses was moved to "core" subtree
from "tools", etc.

The core now also creates a directory in debugfs where all its control files
should reside, including the file created by the section subsystem.