core: Made "Sections" subsystem more flexible
kedr_get_sections() can now be used for more than one module at a time. The
list of sections is no longer a "global state" of the subsystem and the
execution of kedr_get_sections() is now properly synchronized.

If we would like to process several target modules simultaneously some time
in the future, we will be able to call kedr_get_sections() for each of these
(with different 'sections' lists as the argument).