core: Added top-level interface for instrumentation; revisited tests, etc.
1. Added struct kedr_i13n, the structure to define "instrumentation
objects". Each instrumentation object is created for a target module. The
goal is to reimplement the instrumentation subsystem so that it had no
mutable global state. This would allow using this subsystem for several
modules in parallel.

2. The path to the directory where user-mode helpers should be located is
now a parameter of the module (default is the directory where these scripts
are to be installed). This is mainly valuable for testing, because the tests
are run before the package is installed. The tests have been revisited to
use this.

3. Empty value of "target_name" parameter is no longer allowed.