core: Implemented support for hooks
Hooks are the callback functions to be called at the different phases of the
instrumentation process. Currently, one hook is supported, it is called
after the function lookup has been completed successfully. More hooks are to
be added along with the appropriate subsystems (IR handling, etc.).

One of the main reasons for such hooks is to simplify testing and debugging
of the core. In the future, the hooks could probably be used for other
purposes as well.