core: We must not exclude "init_module" and "cleanup_module"
"init_module" and "cleanup_module" are not always aliases to other
functions. There are modules in the kernel that do not use module_init() and
module_exit() and just define "init_module" and "cleanup_module" explicitly.
Therefore, we should not skip these symbols when looking for the functions
to be instrumented.