core: Implemented building of the intermediate representation IR of the code
IR is created for an instrumentable function in the target module when this
function is processed. Currently, the IR is a list, the nodes corresponding
to the machine instructions.

Among other things, the code is split into blocks at this stage. The part of
the information about memory access events in each block that is already
known is recorded in "block_info" structures. These structures are similar
to "TRACE passports" from TSan by their role.