devel: Added "asm-offsets" example
An example to demonstrate how to create a list of definitions of the field
offsets into structures in a header file. The file can then be used even
from the assembly code.

The problem is, the assembly code (neither normal nor inline) cannot
currently use constants defined using 'offsetof' or the like as the offsets
of fields in the structures. For example, the following seems impossible if
<offset_of_some_field> is not a literal integer constant and is not derived
from such constants via arithmetic operations:

mov %rax, <offset_of_some_field>(%rbx)

As a workaround, a header file is generated that contains the needed offsets
as such literal constants. The list of the definitions to generate is in
kedr_asm_offsets.c.