core API: on_memory_event() will be called for all events in a block
on_memory_event() will be called for all events that could happen in a
block, even for those that have not actually happened. In the latter case,
0 will be passed as 'addr' to on_memory_event().

Consistent with that, begin_memory_events() will receive the number of
events that could possibly happen in the block as 'num_events' rather than
the number of the actually happened events as it was supposed before.