core: Made call/jump thunks more Ftrace-friendly
The special functions from Ftrace (mcount(), ...) may preserve more
registers than it is required for the ordinary functions. The compiler seems
to take this into account and not to save/restore these registers before and
after 'call mcount', respectively

The thunks our system uses when handling function calls now restore all
the scratch registers (compared to %rax and %rdx only) when Ftrace is
enabled in the kernel configuration. This allows to avoid problems in case
some post handler uses the registers that the target function was meant to
preserve.

It could be probably more efficient to process only the actual calls to
Ftrace functions in a different way than the other ones. But it is harder to
implement and maintain than what I did in this commit.