functions: Added the mechanism to handle particular function calls
"kedr_func_drd" module should handle the function calls that can be
necessary for data race detection (lock/unlock, alloc/free, ...). The code
of the handlers is created using the templates by "kedr_gen" tool, similar
to what is used in KEDR.

For the present, only a couple of dozens of functions are tracked this way.
Support for more functions should be added in the future.