core: Fixed: different assumptions on the saved value of %rsp/%esp
The thunks used in the processing ofthe function calls and the function
handling subsystem had different assumptions on what value the slot for
%rsp/%esp in the local storage should contain when calling the pre-/post-
handlers.

All components now assume that the value should be as it is before the call
to the target function in the original code. That is, if the target receives
some of its arguments on stack, the saved value will point to the first of
such arguments.