devel: Added utility scripts to be used when working with TSan
"convert_trace_to_tsan.awk" converts the trace prepared by the test reporter
to the format that the offline version of ThreadSanitizer can process.

"symbolize_tsan_report.awk" converts the raw addresses in the stack frames
of the TSan report to to form "symbol+0xoffset". The rest of the report
remains unchanged.