func: Removed registration of accesses to the user-space memory
The system may pass the user-space buffers with the same addresses to the
functions of the target module. That is, even if read() callback has been
called from several threads concurrently, it may have received the same
address of the buffer each time even if the buffers themselves are
different.

To avoid false positives on such accesses, I have removed registering
accesses to the user-space memory from the functions like copy_*_user, etc.

In the future, the addresses of memory areas read and written from the
target module directly should also be checked before reporting to filter out
accesses to the user-space memory. Such situations may arise when something
like get_user/put_user is used, for example.