core: Implemented runtime address filtering
If 'process_stack_accesses' is 0, not only %sp-based memory accesses will
be ignored, but the addresses of accessed memory will also be filtered in
runtime. For the addresses of stack areas, no memory events will be reported
in this case.

Similar runtime filtering has also been implemented for the addresses in the
user space (controlled by 'process_um_accesses' parameter).