core, func: Added support for "memory acquired / released" annotations
"memory acquired / released" annotations allow to specify the "lifetime" of
a memory block the target module did not allocate but obtained from some
other component of the kernel.
For example, it can be convenient to annotate 'struct file' passed to
the file operation callbacks as if it was allocated in open() and
deallocated in release(). This allows to avoid confusion in case the
memory occupied by the 'struct file' is reused after the structure has
been destroyed.