core: Made sampling-related counters available for reading from debugfs
The total number of common blocks containing potential memory accesses as
well as the number of blocks skipped due to sampling can now be read from
the files "blocks_total" and "blocks_skipped" in the core's directory in
debugfs.

These counters are not expected to be absolutely precise (they are updated
in a racy way among other things). They can be used for collecting
statistics and for debugging.