func: Fixed: handle NULL source addresses properly for *dup*-functions
Although it is probably not guaranteed it will always work, kstrdup() can
be called with NULL as the address of the source string. As of kernel
version 3.4, ext4_fill_super(), for example, relies on that.

Before this fix, the handlers for *dup*-functions expected the pointer to
the source memory area to be non-NULL, which led to crashes sometimes. I
have adjusted the handlers, they should now process such situations
correctly.