core, functions: Added several utilities for the function handling
1. Added support for pool of IDs unique during a session with the target
module. These IDs can be obtained with kedr_get_unique_id() and used, for
example, as the identifiers for the groups of the happens-before arcs, etc.

2. Added preliminary implementation of the mechanism that creates a happens-
before arc between the end of the module's init function and the start of
its exit function. It might be replaced later with an implementation based
on KEDR-COI or whatever.

3. Added on_target_loaded/on_target_about_to_unload notifications for the
function handling subsystem. It might need these to obtain or update the
relation IDs, provide the connection points to KEDR-COI, etc.