core, func: Added call replacement API for the plugins
The function handling subsystem now has the API for plugins. Currently, at
most one plugin can be used at a time. A plugin provides a replacement table
for the functions that can be called by the target module. The function
handling subsystem ensures that the appropriate calls are replaced according
to this table.

This is similar to the interaction between KEDR core and payload modules but
in KernelStrider, it seems to be more of a prototype at this stage rather
than a final API.

Among other things, a plugin relying the mentioned API could use KEDR-COI
or whatever else to establish happens-before links between different parts
of the target module (e.g. between the registration of a callback and the
execution of that callback). The current set of event handlers, TID, etc.
can be obtained in the plugin via the functions exported by the core.