devel: Prepared an application to resolve symbols in TSan reports
'symbolize_tsan_report' takes the report produced by ThreadSanitizer and a
file with symbol names and addresses. Then it replaces the raw addresses of
the call stack items in the report with appropriate symbol names and the
offsets into these symbols. The resulting report is output to stdout, the
original one remains unchanged.

The application replaces symbolize_tsan_report.awk, a terribly slow script
I have been using for that purpose before. The processing of the reports
should be much faster now.