core: Implemented the basis for providing data about the functions in runtime
The internal function entry handler now gets the structure where various
data needed to process the function can be stored rather than just the
address or the original function. The address of the structure is passes to
the handler as its only argument and is stored in the local storage too.

This may allow, for example, to implement the support for additional pre-
and post-handlers for the functions, the handlers that will be called no
matter if the corresponding functions are called from the target module or
from some other component. This can be useful when processing callback
functions.