core: Added the means to lookup information about the function in runtime
When the target module is loaded, kedr_find_func_info() can be used to find
the information about a function at the given start address (the address of
its original, i.e., not instrumented instance). Currently, the most
important parts of this information are pre- and post- handlers.

For example, if some kernel-mode component needs to track the execution of
a callback function provided by the target module, it may perform the
following steps:

1. Find the address of that function (may be by intercepting the
registration of that callback).

2. Call kedr_find_func_info() (provided by the core) with that address to
find the func_info object for that callback.

3. Set pre- and/or post- handlers in that object. The handlers will be
called no matter whether the callback is called from the target module or
from any other part of the kernel.

Note that currently only one pre handler and one post handler are supported.