core: Added the means to get the arguments in the handlers for the callbacks
The prologue and the epilogue code inserted by the core into the
instrumented instances have been modified to save the registers needed to
obtain the arguments and the return value of a function. Now KEDR_LS_*() can
be used in the handlers for the callback functions too, no matter how these
functions have been called (by the target module itself or by another part
of the kernel).

Previously it was possible to obtain these values only in the handlers for
the functions called from the target module itself. That was not enough
for the processing of the calls to the callbacks.