func: Added processing of the operations with character devices
The registration and deregistration of the character devices as well as of
the corresponding file operations can now be processed. The calls to some
of the callbacks (open, release, read, write, llseek) are handled too.

The handlers create appropriate happens-before arcs and generate other
events to record the following relations:

- each of the callbacks starts not earlier than the registration of the
callbacks starts for the given character device;

- each of the callbacks end no later than the deregistration of the
callbacks ends for the corresponding device and strictly before the exit
function of the target module starts;

- 'struct file' passed to each of the callbacks "appears" on entry to the
callback and "disappears" on exit (as if it were allocated and deallocated
this way). This is a simplification but it should be enough for the present.