core: Fixed: function names should be copied
As it turns out, it is unsafe to store a pointer to the name of a function
(a string in the string table of the target module) and expect the pointer
to be valid in runtime. The module loader may change the string table in the
target's memory image after the target has completed its initialization.

So, as far as kedr_ifunc structure is concerned, we need to prepare a copy
of the function name during instrumentation and work only with that copy
later.