core: Added the means to resolve the addresses of the instrumented code
Given an address of a location in the instrumented code of the target module
(IP, PC, ...), it is now possible to quickly determine the function the
address belongs to. This may, for example, simplify analysis of the oopses
in the instrumented target modules. The IP recorded in the system log in
such situations will point to the instrumented code the kernel symbol
resolution system does not know of. The means added by this commit allow the
user to request the needed information from KernelStrider core.

The core provides 3 more files in its directory in debugfs. To resolve an
address, write it (a hex value) to "i_addr" file then read the contents of
"func_name" and "func_i_start" files. If the address belongs to an
instrumented function, "func_name" will contain the name of the function,
"func_i_start" - the start address of the instrumented instance of the
function. If the IP is unknown, "func_name" will contain an empty string and
"func_i_start" - 0x0.

Example.
Suppose the target module "my_cool_fs" crashed during the analysis. From the
system log one may be able to retrieve the IP of the corresponding
instruction:

   IP: [<ffffffff8124dce0>]

Let's find the corresponding function (root privileges are needed):

   # echo ffffffff8124dce0 > /sys/kernel/debug/kedr_mem_core/i_addr
   # cat /sys/kernel/debug/kedr_mem_core/func_name
   my_cool_fs_get_sb
   # cat /sys/kernel/debug/kedr_mem_core/func_i_start
   0xffffffff8124dc10

The address of the original instance of the function is usually not needed
in such cases (it did not execute anyway, except a jump to the instrumented
function). Still one can obtain it from /proc/kallsyms, if necessary.