util: Added an output system for testing and experiments
This output system provides less functionality than the full-fledged one
will (see sources/output/ subdirectory) but it can be convenient to use for
testing and experimenting with the real-world target kernel modules.

The system is based on the idea that Perf uses for data output: a ring
buffer maintained by the kernel-mode part of the system is mapped (mmap())
to the memory of the user space application. The latter saves the data to a
file. The implementation is a great deal less optimized but simpler than in
Perf (no per-cpu buffers, etc.). It still remains to be seen whether the
overhead of our system is significant.