util: Added utilities convenient when working with ThreadSanitizer
- kedr_show_* scripts display the names and addresses of the ELF sections
(or both of the sections and of the symbols) belonging to the target module.
The scripts may be used when the target module is loaded. The results can
be then fed to "kedr_symbolize_tsan_report".

- kedr_symbolize_tsan_report processes the report produced by
ThreadSanitizer and resolves PC (IP) information there using the list of
symbol names and addresses prepared by kedr_show_* scripts before.