core: Revisited the expected traces in the tests after format change
The commit 4796dd200db943e36f876e7029552212e5bbdf33 changed the way how data
are output if %ps and %pf specifiers are used in printk and the like.
I have revisited the test scripts and the expected trace files for the tests
to be able to work both with that change and without it.