func: Added the means to get the arguments of the functions with varargs
On 32-bit x86 systems, the kernel functions with variable arguments lists
(e.g. sprintf()) get all their arguments on the stack. The ABI does not
specify that though, so this may be a GCC-specific feature or something like
that. As a result, KEDR_LS_ARGn() cannot be used for such functions.

On 64-bit x86 systems, the ABI imposes similar rules for the functions
with variable argument lists as for the ordinary functions: normally, the
first 6 arguments are passed via registers, the rest is on the stack.

A set of KEDR_LS_ARGn_VA() macros has been provided to access the arguments
of the functions with variable argument lists in a uniform way in the pre
and post handlers.

The tests checking that the arguments are retrieved correctly this way.

Note that the functions having an argument of type 'va_list' (e.g. vsprintf)
are affected by the same rules as the functions without variable argument
lists. KEDR_LS_ARGn() should be used for these functions rather than
KEDR_LS_ARGn_VA().