core, func: Address filtering is now performed for all read/write operations
The core now performs address filtering in kedr_eh_on_memory_event(). This
way, not only the events detected by the core itself but also those
reported by the function handling subsystem will be processed.

As a result, read and write events for the stack-based and user-space-based
objects will not be reported (if the appropriate parameters of the core are
set), no matter if they come from a plain memory operation or from something
like kstrtoul(), for example.

Function groups "dup" and "strings" have been revisited to make use of
kedr_eh_on_memory_event() (and other kedr_eh_*() functions as well) instead
of direct calls to the handlers.