examples: Added "buggy01_annotated" and "test_buggy01"; revisited the tests
"test_buggy01" is a simple application that reads from the file in debugfs
maintained by "buggy01" module. The application currently does that from 2
processes.

"buggy01_annotated" is the same as "buggy01" except happens-before arcs in
it are annotated as well as the "allocation" and "deallocation" of the
memory pointed to by 'filp' in the callbacks.