tests: Fixed: the tests should use their own trace sender module
Before this commit, the tests for the CTF-based output subsystem used
kedr_trace_sender module which, in turn, used the symbols exported by
kedr_mem_core module. The tests, however, do not use the latter, but rather
a stub that exports the same symbols as kedr_mem_core among other things.

So, in some cases, kedr_trace_sender module failed to load in trace_sender.*
tests because the checksums of the symbols it needed diffed from what the core
stub provided.

The problem was first found in ROSA Marathon 2012 x64 with an experimental
"nrjQL" kernel (version 3.6.6).

This commit fixes the problem/