func: Fixed processing of handler arrays provided by FH plugins
A bug has been fixed that prevented using 2 or more FH plugins with the
function handlers at the same time. A sanity check was made in a wrong place
in the code.

A test for such situations has been added: the FH plugins "common" and
"cdev" are used there in a single analysis session.