utils: Revisited the simple output subsystem to work with multiple targets
The simple output subsystem is now able to process the events from several
target modules.

Session start/end events are now handled; the user-space part now exits on
session end rather than on target unload, etc.

Note that the format of "target load" and "target unload" events has
changed: the name of the target is now passed there along with the addresses
and sizes of the "init" and "core" areas rather than the address of struct
module. The information about "init" and "core" areas of the target will
make it possible to determine the start addresses of the target's sections
(not implemented yet) without using kedr_show_target_sections.sh script.

The existing tests have been updated according to the changes in the
subsystem. A new test has been added that involves 3 target modules to check
that the subsystem handles such situations properly.