build: tsan_report_addr2line should always be built
Without tsan_report_addr2line or a similar tool, it is very difficult to analyze
the collected data. I suppose it makes no harm to build this tool
unconditionally.

As a consequence, development files for libelf are now always required for the
build.

The option "WITH_REPORT_ADDR2LINE" has been removed as well.