util: Added an "all-in-one" trace processing tool ("tsan_process_trace")
"tsan_process_trace" tool simplifies the processing of the trace collected
in the kernel, especially if there are two or more target modules.

The tool takes the trace and the binaries of the involved kernel modules,
locates ThreadSanitizer offline, used it to analyze the (appropriately
converted) trace and outputs the report.

If the modules' binaries contain debug info, the source lines will be
specified in the call stack items in the report, otherwise - ELF sections
and offsets.

"tsan_process_trace" relies on libelf and libdw from elfutils to deal with
the debug info. Neither addr2line nor eu-addr2line are used here.

To get more details about the usage of the tool, one may execute
    tsan_process_trace --help

Currently, the tool handles only the traces collected using the simple
output subsystem. In the future, it may be enhanced to be able to process
the traces from the "professional" output subsystem as well.