tests: Added a test for detection of common kinds of races ("bug_bench")
In this test, everything is performed in almost the same way as a user would do
it.

The modules "kedr_sample_target" and "buggy01" are used here. Both contain racy
code, so the test checks if our system working with ThreadSanitizer detects all
of them.

"tsan_process_trace" tool is used (and therefore checked) in this test.

Note that the test is enabled only if ThreadSanitizer offline is found somewhere
in $PATH during the configuration (cmake) of the package. The configuration
system first looks for (amd64|x86)-linux-debug-ts_offline, then - for
ts_offline.