core: Fixed handling of targets that have init area with no code
Before this commit, the core handled the target modules have init area
with no code incorrectly. It checked that struct module::module_init was not
NULL, assumed there is code in that area and tried to handle that. Same for
core area.

This resulted in "Failed to allocate memory for fallback functions" errors.

"crc*" modules are the example of such modules. They seem to have data but
no code in their init areas.

This commit fixes the problem. A regression test has also been added here.