core: Use in_irq() rather than in_interrupt()
It turns out, it is not really needed to process softirq context in a different
way compared to the process context when assigning thread IDs and dealing with
stack. The code running in softirq context actually runs in the dedicated
threads and can be handled as such.

In addition, in_interrupt() is unreliable: it may return non-zero in the process
context if bottom halves are disabled, e.g., if a spinlock is locked with
spin_lock_bh(). This lead to problems with the drivers for wireless networking,
for example, where a LOCK event for such spinlock was assigned to one thread and
UNLOCK - for another one (a spurious interrupt handling thread). This did not
make ThreadSanitizer very happy.

Handlers for hardware interrupts, however, still need to be processed as before.