util: On target load, size of the code in "init" and "core" is now reported
Before this commit, the full size of "init" and "core" memory areas was reported
by the simple output subsystem. This lead to more problems with the target
modules that have no code in their "init" areas: tsan_process_trace failed to
determine that the initialization of such modules had completed and the memory
from their "init" areas was available for other modules again.

Got around this by reporting only the size of the code in these areas rather
than the full size. This information is only used to find the addresses of the
code sections, so the full sizes are probably not needed anyway.

kedr_test_reporter has also been updated this way, for consistency.