core: Revisited thread handling and event sampling subsystems
The thread handling subsystem has been revisited to detect started and ended
threads more reliably.

1.
"Thread start" and "thread end" events are now generated by the core. The former
is generated for a thread when it is about to begin to execute the code of the
target modules the first time during the session. The latter is generated when
the core knows for sure a thread that executed some code of the targets has
ended.

"Thread start" events are generated for each thread that executes the code of
the targets.

"Thread end" events are not necessarily generated for all such threads. If there
were no such event for a thread, one can assume the thread remained live when
the analysis session ended.

2.
For an ordinary thread (i.e. not an IRQ), the core uses the address of its
task_struct as the ID. However, once a thread has ended, the memory occupied by
its task_struct is freed and is often given to another thread's task_struct.
This way, our system was unable to distinguish these two threads and could miss
some races as a result.

The new trace handling system detects such task_struct reuse. Besides that, the
core scans the table of threads periodically to detect the ended ones and
process them.

3.
Hardirqs and softirqs are again handled the same way. It seems, softirqs are not
guaranteed to be executed in their dedicated threads and may be handled like
hardirqs. A more reliable replacement to in_interrupt() is now used to avoid
confusion when the code executes in the process context but with the bottom
halves disabled.

4.
For IRQ handling routines, the CPU id is used as a thread ID as before.

5.
The event sampling subsystem has been rewritten to take advantage of the new
thread handling mechanism and it has been simplified.

The output subsystems and the user-space trace processing tools have not been
updated yet to take advantage of the new thread handling mechanism. This will be
done in the future.